Create Stopwatch code
This is a simple implementation of a stopwatch class in JavaScript. You can create an instance of the class and call the start(), stop(), and getElapsedTime() methods to start, stop, and get the elapsed time of the stopwatch, respectively.
The class is using Date() object to get the current date and time, and also maintain the state of the stopwatch by using the variable running.